SwiftCart/requirements.txt
Django==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
pytest==7.3.0
pytest-django==4.5.0
requests==2.28.0
